created regions_hash to process regions inputs for scrapper
